Revert "Fix wrong ColorBuffer refcount initialisation"
This reverts commit 73a77025a3306c7ca239e2b733a5fc41a0e5ca23.

This fix corrects memory leaks on 4.1.1 and 4.2.2, but was in reality a
workaround for a fix that was missing in these versions. it introduces
graphical errors (black textures) on 4.3 and 4.4.2.
These 2 versions contain the right fix in the opengl system libs that must be
reuse on 4.1.1 and 4.2.2 version:

+ stream->flush();

in renderControl_enc.cpp, at the end of rcCloseColorBuffer_enc() function.